open-app-filter: fix the issue where the kernel repeatedly tries to automatically load the module when the oaf module is not loaded
When the oaf kernel module is not loaded, this process will appear once every second `/sbin/modprobe -q -- net-pf-16-proto-29`.

This is because the kernel is automatically looking for a suitable kernel module, but we don’t want this, so check that the oaf kernel module is loaded before opening netlink.